fix bugs in database setup

+++ db/schema/01_users.sql
@@ -3,7 +3,7 @@ CREATE TABLE users (
   id SERIAL PRIMARY KEY NOT NULL,
   name VARCHAR(255) NOT NULL,
   email VARCHAR(255) NOT NULL,
-  phone_number INTEGER NOT NULL,
+  phone_number BIGINT NOT NULL,
   password VARCHAR(255),
   is_employee BOOLEAN DEFAULT FALSE
 );
